feat: add css reset
I want to control everything about the layout of this application therefore a css reset helps me to get rid of the default styling.

feat: add styling for general information inputs

These newly added styles layout out the labels and inputs within the form in a neat manner making the user experience better and flowing to the next section.