feat: add logic to render general input / output
Depending on the status state variable in this component it'll either render an input version of General or an output.
The input version allows users to input data into fields which save to the App component's person state variable when the user makes changes in the input fields. These input fields load the saved person state data (controlled elements) when rendered.
The output version of General displays the data in a presentational CV layout, which users can print out and use.