feat: add state to Education component
This state will be used to keep status on if the user is editing their education or have submitted their education.

styles: import style file to be used

feat: add conditional return of component

The condition resets on the status state variables value and depending on if it is 'edit' or 'submit' the user will be rendered input fields or saved data in a presentable fashion, which I have not finished but have left comments to remind myself next. This also includes using another component for the input which I've named EducationInput