fix: rewrite what is rendered in situations
Now this component renders with initial default values, empty strings.
The component now attempts to render the correct amount of EducationInput components depending on the number of objects in the education array of person.

feat: add functionality to add button

This buttons onClick now triggers creation of a new education array object and thus triggers a render using the above rewritten logic to generate an EducationInput component.

style: add classNames to generated elements

fix: edit dates output to be in M/D/Y format

This is easier, at least for me to read the order of the date this way.

feat: replaced comments with statements that evaluate to data

This is in regarding the pulling stored data from person and inserting it into input fields to display.