fix: rewrite component
Added a helper function that will be triggered onChange in any of the input fields, which uses the index prop to pick out the specific education object this component is tied to and change stored values in that object.

By changing App person state variable to have an initial object with properties we can now load those empty properties instead of using a condition to load an empty string when that object/property is not found, and avoid any errors related to controlled inputs switching to or from uncontrolled.