feat: add validityChecker function and usage
This function will be invoked whenever there is entry in a required input field at which point it will check if that input is valid against any input attribute types or patterns, and if not valid will apply styling and insert an error message below the input element until the entry is valid.

fix: change asterisk to required fields

Prior to this change the asterisk was only on optional fields, which didn't make sense so I switched its usage.

feat: add required attribute to select inputs

Some inputs are required and pattern attribute was also added to help validate user entry data.

fix: change disclaimer to placeholder text for skills input

feat: add validity check of Submit button

This button when clicked prior to these changes would change the status state variable, but now it checks if the required input elements entry is valid before it makes that change. If not valid it displays an error message to the user and applies error highlighting to the outline of those inputs.