style: add rule for add education error message
The message that is created and displayed it a span so by default it sits inbetween the buttons of add another, which I'll rename next, and submit, which is ugly. So I've gone ahead and changed that span to have a block display so that it takes up an entire horizontal line of content by itself.