docs: remove or rewrite comments
These are for explaining what I'm doing if complex, and contain notes for code I need to write in the future

fix: change argument to key & index EducationInput

These arguments are index values of the new sliced array, but we need index of the elements in the original array, so we'll increment these arguments by one because we're using a slice of the original array minus the first education object that gets rendered by default.

feat: add a limit of ten EducationInput objects

Otherwise a user could just keep adding more objects. There is right now a problem, if a user adds more than they need or want, there is no way of getting rid of them. So I need to add a delete button for that. In addition to the limit I've added an error message that pops up if a user reaches that limit of EducationInput components.

feat: add validity checks for EducationInput

These validity checks on input elements disallow the user to submit incomplete education data to the submit result which would create empty ugly list items.. By adding a future delete button we'll give users the ability to remove falsely made education objects that make their CV look ugly. Applied styling for when the required education inputs entry is not valid, and later will add messages under each input to describe to users the problems of their input.

fix: remove some education property checks

When a user submits there is a condition that checks if there are education objects in person, before I added required to input elements this condition also checked for an institutionName length greater than one before it rendered which isn't needed any longer because we have added validation checks and requirements to input fields including institutionName.

fix: change how dates are rendered

Before this change if a starting or ending date value was empty then it would render empty before or after /'s in the format of the output when rendering submitted data. Now startDate is a required field and endDate is optional and when not entered 'present' will be rendered instead of an empty string.