feat: add asterisk explainer/disclaimer
This tells users that an input field wrapped by a label with an asterisk in its text content denotes that field is required to have input.

feat: add asterisks for content of required inputs

This obviously tells the user that this input is required.

feat: add form

Before the input elements were just existing within the component instances, even though we don't use submit for anything because our buttons are buttons and not submit type I've gone ahead and wrapped the labels and inputs with the form.

feat: add delete button

This button allows users to delete EducationInput components, therefore removing not only the educations they've added but potentially the default education object as well, which is ok because we have dynamic rendering in the submitted form of our application depending on what they submit it will look presentable.

fix: change date descriptors startDate and endDate

These names we're changing to already existed in the education objects of person by default so I'll go ahead and use them instead.