feat: add functions file
This file stores the functions used by components across the application.
changeArrayObjectsValue is used to save state of the Person when needing to change a value of a property of an element in an array of a category of Person, like education.title. This will be used in the professional section as well
validityChecker is very important as it's used in all of the input components to check if user data input is valid for each input field that is required.
saveInput contains the most basic property value change functionality but is still cleaner to call than the three lines it contains.
verifyInputs is invoked when the submit button of a component is clicked and checks if the inputs of that component type are valid, if not applies styling and an error message below the submit button.