fix: change classes to id's for elements
With the changes to onChange functions being moved to a module elements require more uniqueness and changing from classes to id's gives them that guaranteed unique value.

feat: pass functions into onChange to be invoked

One of these functions existed in this component prior to moving it and others to a module, but now these get imported from the module to save us from dry code

fix: remove unnecessary prop of key

key is used in the parent of EducationInput which is a div, so it's not needed to be passed to EducationInput any longer.

feat: add invocation of verifyInputs

This is used when users click the 'submit' button in an EducationInput instance, it's used to communicate to users that their input  is insufficient.