fix: change classes to id's for elements
With the changes to onChange functions being moved to a module elements require more uniqueness and changing from classes to id's gives them that guaranteed unique value.

feat: pass functions into onChange to be invoked

One of these functions existed in this component prior to moving it and others to a module, but now these get imported from the module to save us from dry code. These functions when invoked by onChange save input to state and check for validity of the user data against the input elements requirements, and communicate to the user through messages and styling when their input data is insufficient.