fix: change condition to generate first education
The condition I used to render the first education object was checking if the length was === 1, when it should have been checking for when the length > 1. So it was generating one instance of EducationInput but when there were two or more education objects then it would stop rendering the first instance.