docs: add import and export statements
Imports include methods from a project module that we'll use to validation and verify inputs and to save input to our state variable of person. And of course exporting this component so it can be used in the App Component which is the root component of our application.

feat: create a variable with useState

This variable will track if our component is in 'edit' or 'submit' mode/value, and that value will be used conditionally to render input components or output of data of person.education objects.

feat: add conditions to render dynamic content

These conditions check the number of person.professional objects, and depending on that the user will see input fields or a statement alerting them that there are no professional objects in person and in either case gives the user two buttons, one to add professional objects to person, and the other to render their professional objects into output or to hide the section in the output / 'submit' state so that there isn't an empty Professional section rendered on their CV.

feat: add starting component invocations

These don't include props yet, they are just boilerplate using the child components name.