feat: add function for component creation
feat: give component person and index prop imports

feat: have the component render saved data

This component pulls data from the person state variable, specifically it's professional property which is an array of objects and generates elements for each piece of that data of the object. Some data requires output to be dynamic, for example if a professional object has a start date but not an end date then rather than display an empty string it'll show "Present" as it's presumed that the user is still working at that job. Another example would be responsibilities are generated into a list iteratively.