fix: remove unnecessary pattern attribute
The dateStarted input element had a pattern attribute which was not actually being applied because for some reason date type inputs validate on full entry, aka when a user has filled in a value for each section; month, day, and year. Now validation just requires input which again is input into each section of date.