fix: change ul to section, li's to sections
When generating container elements for multiple education objects I've changed generating them into list items within a unordered list because I thought it would be better to have semantic attachment, each of the major sections of this application take in and render data in sections that are related, so I feel there's a need to keep that semantic connection.

fix: change h3 to an h2

This rendered section resembling a components data didn't have an h2 and to keep that hierarchy I've changed the h3 to an h2.

fix: wrap date creation in span with 'to'

This was changed because how we use flex to layout it out results in unwanted sections being placed apart from each other.

 fix: add class to edit button

This is needed so we can hide the edit button on the print media type so that the user can print a somewhat useable cv.