feat: change p to h3 for title
I think a persons title which is there position they're looking for/have worked in the past is very important to stand out so have changed it to be an h3, although I may change this later because it doesn't necessarily make sense in the hierarchy of heading elements, but does imply semantic importance.

 feat: remove class from description

Previously this description was in a larger font-size which I've changed my mind on how important that is, thus changing back to default p element font size.