feat: add addProfession import
This function will allow the user to create more professional experience objects so they can add more jobs they've worked in the past. By putting this in a module I'm saving the clutter to a more appropriate place.

refactor: rewrite render when status === 'edit'

Upon testing I noticed that the elements being rendered weren't quite what I needed and that would be fixed by refactoring the entire return.

feat: add logic for add professional experience

This button had no code tied to it and was not attempting anything when clicked, now creates another object in person.professional array.

fix: remove setPerson invocation from submit

This was an accidental addition, verifyInputs calls setPerson, which is bad because that function is doing too much but anyways.

refactor: rewrite render when status === 'submit'

Just like the 'edit' version, the output was wrong and I needed to fix it, actually this one was worse, there were two "Experience" headings that were rendering and that's what alerted me to needing to rewrite the section.