feat: add import statements for three functions
These functions are used to change responsibility values, delete a responsibility, and delete a professional work experience from person.

feat: add logic to render necessary inputs

I added a lot of code that renders the input sections of this component, sadly I was hyper focused on writing working code and not committing my changes so this includes about 180 lines.
This component will be called for each professional array element in person.
I could have made a responsibilityInput and output but didn't, may do that in the future, but stuck with just making dynamic inputs for each saved responsibility.
There's too much to mention really and I'm sorry for not committing more often.