feat: change h3 to h2
To keep more of a hierarchy of headings I've made this change, this one specifically for the company name the user has worked for.

feat: wrap date 'to' in a span

This is to give that 'to' styling so it sticks out from the date values.

feat: add 'responsibilities' above rendered list

Although I think I should have used a heading, which I actually gave it a className that included 'heading' in it, I wanted to change the styling for the word to stick out.

feat: add class to ul

This ul needed styling for layout, because it was ugly without it.