style: change add responsibility button to block
This allows it to be on a line of its own and not clutter that line with multiple buttons.

style: apply grid to forms

This allows the forms to have a structured layout for labels and buttons, which looks nicer than flow or flex layouts.

style: change delete responsibility button width

Just like previously changed buttons this was to shrink the button to have a width of its content, nothing larger like grid forcing it to be 100% of what was being offered to it.

style: change width & margin-bottom of  inputs

By changing width to fit-content, I've limited the amount of horizontal space which makes these professional component renderings to only be as wide as their content needs to be and not the entire page which was very ugly.
By adding margin-bottom we separate each of these rendered components from each other, the horizontal gap helps to determine there's a difference between each component rendered event hough we also use a horizontal line break to imply this separation.

style: apply vertical margin to hr's

This vertical margin helps some more with separation between each rendered professional input component.

style: apply flex to labels

This is for laying out the label text content and the input element field neatly, which could have also been done with grid but this was faster.
Also applied some bottom margin to space label and inputs away from each other so a user has a less likely chance to mis click into the wrong input.

style: set width of input fields

I've settled on a specific width as I think it looks neat, however I just noticed that the skills input field is not quite placed in line with the input fields preceding it, which is ok because I'm going to refactor that input into many inputs and I'm sure I'll need to restyle them later to match how the responsibilities section in professional experience has come out.