style: add styles to professional-output
To match previous sections I've applied grid with some margin-top to separate it from the preceding section, and aligned the text content to the center to make the output more easy to read.

style: style the rendered sections

This is primarily to set them apart from their containers heading and to match placement of education objects, where there's a position name and then dates attended and then below the name is the position at the company just like degree/course title completed

style: style the position name and date

This is so that we get the company name and dates employed to be on the same line.

style: change font-style & alignment of  position

I wanted to set it apart from regular text so used italic font-style which might not be good and can change it later.

style: position and bold'en 'responsibilities:'

This is more of a mini heading that just some text as it sits atop rendered list items.

style: apply styling to rendered ul

I've tried to change how the responsibilities are laid out but am still not happy with it and will have to change this later.

style: remove default padding from ul

I thought my css reset would have removed this but here we're removing it for total control.

style: apply style to 'to' of employed dates

Just like other sections I wanted the 'to' to stick out from the dates because I found it annoying to look at the dates as a unit and couldn't tell quickly there were two dates for starting and ending.

style: add media print type query

These are to hide the edit button, the section heading when there are no education objects, and the border around the section when printing the page so that the user can actually use the output as their cv.