feat: add deleteResponsibility
This function will be used to delete single responsibilities and will be tied to 'delete' buttons next to responsibilities.

feat: add deleteExperience

This function will be tied to the 'delete' that sits within each professionalInput component instance, and of course deletes that object from the professional array.

feat: add addProfession

This function is tied to the "Add Another Professional Experience" button and will create a starting professional object in the person.professional array