feat: add addEducation import and invocation
This function now holds the logic to add a new education, or if there are ten education objects alert the person that they have too many, and ten is probably five more than too many as it is.
Moving this function saves us some clutter.

fix: remove unnecessary render of object

Removed component render that could have been done in the iterative way that followed it. And edited that iterative technique to include it of course.