refactor: move General input renderings
This logic has all been moved to GeneralInput, including the importing of the functions used in that form of inputs to validate and generate error messages to users. The setStatus state variable function has been passed as a prop to it so it can change 'status' just as it did before.
Next I'll create a skills component to replace the input element that asks users to enter values comma space separated as that is terrible to ask users to do.