feat: add id to form
This id may be needed to style elements in components that exist within the GeneralInput component.

refactor: rewrite rendering logic for skills

The code that generated skill input elements and buttons and listeners was moved to a new component and now called iteratively as to generate a GeneralInputSkill per skill saved to person.skills. This change of course means less messy code in GeneralInput.