refactor: move input styles to specific file
I just wanted to be more specific with where styles are so created this GeneralInput.css file to store styles when GeneralInput is in 'edit' status.

style: add vertical margin on skill inputs

This change helps spread skills away from each other and makes them not so easy to misclick into.