style: move title rule to this file
title id's rule has been moved into both the GeneralInput and GeneralOutput css files because it's displayed on both components and was moved because General.css was divided to the specific component css files.

style: add general input id prefix to label rules

Somewhere these labels are being targeted for styles and I know this because points of specificity elsewhere are causing this rule set to be overwritten. So here I've added a prefix of an id to the GeneralInput form element to give this rule set a higher order and precedence.

style: add margin-bottom to general submit button

This was to give some space between the submit button and the hr separator, it was ugly without the margin.