docs: removed commented out style, add line
I removed an old style statement that was commented out because I found the correct way to style the element and didn't need the old way any longer. I also added a line of space between the media query and the last rule of non media query, so that media query would stick out more.