feat: add default values for startDate
The input elements type was changed from date to month for startDate of education and professional objects. I felt that the specific day is not important to events around employment and education past experiences. Also I gave a default value primarily to give the user an example of input format.