feat: add hr to separate education
This hr is a major separator so users can see that when they click the submit button for a section that they can easily see where that section begins and ends because of that hr line.

feat: add 'Education' to submit button text

I wanted to give the submit button a more specific text content, so that users know what they're submitting for sure.