feat: add hr to separate general from education
This will help the user to understand a little more that when they click the submit general button that it only submits entry in that form and nothing below this separator.

refactor: move output to GeneralOutput component

For organization efforts I've made a GeneralOutput component and passed it person and setStatus so it can handle everything that was generated in the else status==='submit' condition's return body.